chore(structure): split visual and einsum out of notebook, collapse permute trio, release 1.0.0
Extract TensorVisual and the shared rendering substrate into visual.py and the
einsum view into einsum.py, and fold transpose/swapaxes/moveaxis into one
_permute_view helper. Behaviour and the public API are unchanged; notebook.py
drops from ~1388 to ~1000 lines. Bump version to 1.0.0.
